Add Deploy on Paltform.sh button
Since this repo already has Platform.sh configuration files in it (inherited from Ibexa), here's a button to quick-launch on Platform.sh.  I added it where it seemed to make sense in context, but feel free to move it.  The UTM tags are just so we can keep track of which deploy button is which.

Let me know if you've any questions.